Moved the mention processing methods to the already existing MentionUtils class Did not notice it when I made MentionProcessor. Oops
