+++ plugin/src/main/java/at/helpch/chatchat/util/MentionUtils.java
@@ -2,6 +2,9 @@ package at.helpch.chatchat.util;
 
 import at.helpch.chatchat.api.ChatUser;
 import at.helpch.chatchat.api.Format;
+import at.helpch.chatchat.api.User;
+import at.helpch.chatchat.format.PMFormat;
+import java.util.Map;
 import net.kyori.adventure.text.Component;
 import org.intellij.lang.annotations.RegExp;
 import org.jetbrains.annotations.Contract;
@@ -12,6 +15,13 @@ import java.util.regex.MatchResult;
 import java.util.regex.Pattern;
 
 public final class MentionUtils {
+    private static final String MENTION_PERSONAL_PERMISSION = "chatchat.mention.personal";
+    private static final String MENTION_CHANNEL_PERMISSION = "chatchat.mention.channel";
+    private static final String MENTION_PERSONAL_BLOCK_PERMISSION = MENTION_PERSONAL_PERMISSION + ".block";
+    private static final String MENTION_CHANNEL_BLOCK_PERMISSION = MENTION_CHANNEL_PERMISSION + ".block";
+    private static final String MENTION_PERSONAL_BLOCK_OVERRIDE_PERMISSION = MENTION_PERSONAL_BLOCK_PERMISSION + ".override";
+    private static final String MENTION_CHANNEL_BLOCK_OVERRIDE_PERMISSION = MENTION_CHANNEL_BLOCK_PERMISSION + ".override";
+
     private MentionUtils() {
         throw new AssertionError("Util classes are not to be instantiated!");
     }
@@ -67,5 +77,63 @@ public final class MentionUtils {
             ) {
         return replaceMention(prefix + user.player().getName(), component,
                 r -> FormatUtils.parseFormat(format, user.player(), component));
+    }
+
+    public static @NotNull Map.Entry<@NotNull Boolean, @NotNull Component> processChannelMentions(
+        @NotNull final String mentionPrefix,
+        @NotNull final String channelMentionFormat,
+        @NotNull final ChatUser user,
+        @NotNull final User target,
+        @NotNull final Component message
+    ) {
+        if (!user.player().hasPermission(MENTION_CHANNEL_PERMISSION)) {
+            return Map.entry(false, message);
+        }
+
+        if (target instanceof ChatUser) {
+            final var targetChatUser = (ChatUser) target;
+
+            if (targetChatUser.player().hasPermission(MENTION_CHANNEL_BLOCK_PERMISSION) && !user.player().hasPermission(MENTION_CHANNEL_BLOCK_OVERRIDE_PERMISSION)) {
+                return Map.entry(false, message);
+            }
+
+            final var replaced = MentionUtils.replaceMention(
+                mentionPrefix + "(everyone|here|channel)",
+                message,
+                channelMentionFormat);
+
+            return Map.entry(replaced.didReplace(), replaced.component());
+        }
+
+        final var replaced = MentionUtils.replaceMention(
+            mentionPrefix + "(everyone|here|channel)",
+            message,
+            channelMentionFormat);
+
+        return Map.entry(replaced.didReplace(), replaced.component());
+    }
+
+    public static @NotNull Map.Entry<@NotNull Boolean, @NotNull Component> processPersonalMentions(
+        @NotNull final String mentionPrefix,
+        @NotNull final PMFormat mentionFormat,
+        @NotNull final ChatUser user,
+        @NotNull final ChatUser target,
+        @NotNull final Component message
+    ) {
+        if (!user.player().hasPermission(MENTION_PERSONAL_PERMISSION) ||
+            (target.player().hasPermission(MENTION_PERSONAL_BLOCK_PERMISSION) &&
+                !user.player().hasPermission(MENTION_PERSONAL_BLOCK_OVERRIDE_PERMISSION))
+        ) {
+            return Map.entry(false, message);
+        }
+
+        final var replaced = MentionUtils.replaceMention(
+            target,
+            mentionPrefix,
+            message,
+            mentionFormat
+        );
+
+        return Map.entry(replaced.didReplace(), replaced.component());
     }
 }
